+++ Assets/Arkademy/Data/DamageData.cs
@@ -5,14 +5,14 @@ namespace Arkademy.Data
     [Serializable]
     public class DamageData
     {
-        public int amount;
+        public long amount;
         public Character dealer;
 
         public DamageData()
         {
             
         }
-        public DamageData(int newAmount)
+        public DamageData(long newAmount)
         {
             amount = newAmount;
         }
